+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/), and this project
 adheres to [Semantic Versioning](https://semver.org/).
 
+## [Unreleased]
+
+[unreleased]: https://github.com/rogdham/bigxml/compare/v0.7.0...HEAD
+
+### :bug: Fixes
+
+- Fix an edge case when a class decorated with `xml_handle_text` has an attribute
+  decorated with `xml_handle_element`
+
 ## [0.7.0] - 2022-01-15
 
 [0.7.0]: https://github.com/rogdham/bigxml/compare/v0.6.1...v0.7.0